+++ admin/tabs/general/config/ConfigPanelModel.ts
@@ -7,7 +7,7 @@
 import {exportFilenameWithDate} from '@xh/hoist/admin/AdminUtils';
 import {AppModel} from '@xh/hoist/admin/AppModel';
 import * as Col from '@xh/hoist/admin/columns';
-import {div, hbox, hspacer, label, pre} from '@xh/hoist/cmp/layout';
+import {hbox, hspacer} from '@xh/hoist/cmp/layout';
 import {HoistModel, LoadSpec, managed, XH} from '@xh/hoist/core';
 import {FieldSpec} from '@xh/hoist/data';
 import {textArea} from '@xh/hoist/desktop/cmp/input';
@@ -19,7 +19,6 @@ import {
     RestGridModel,
     RestStore
 } from '@xh/hoist/desktop/cmp/rest';
-import {toolbar} from '@xh/hoist/desktop/cmp/toolbar';
 import {action, makeObservable, observable} from '@xh/hoist/mobx';
 import {isNil, truncate} from 'lodash';
 import {DifferModel} from '../../../differ/DifferModel';
@@ -79,13 +78,11 @@ export class ConfigPanelModel extends HoistModel {
                     {...(Col.lastUpdatedBy.field as FieldSpec), editable: false},
                     {
                         name: 'overrideValue',
-                        displayName: 'Override',
+                        displayName: 'Effective Value',
                         typeField: 'valueType',
                         editable: false
-                    },
-                    {name: 'isOverridden', type: 'bool', editable: false}
-                ],
-                processRawData: data => ({...data, isOverridden: !isNil(data.overrideValue)})
+                    }
+                ]
             }),
             actionWarning: {
                 del: records =>
@@ -120,55 +117,16 @@ export class ConfigPanelModel extends HoistModel {
                 {...Col.lastUpdated, hidden}
             ],
             editors: [
-                {
-                    field: 'isOverridden',
-                    formField: {
-                        className: 'xh-config-panel-overridden-field',
-                        label: false,
-                        flex: 1,
-                        inline: false,
-                        minimal: false,
-                        readonlyRenderer: v =>
-                            v
-                                ? toolbar({
-                                      className: 'xh-config-panel-overridden-field__toolbar',
-                                      items: [
-                                          Icon.warning(),
-                                          hspacer(),
-                                          'Overridden by Instance Config'
-                                      ]
-                                  })
-                                : null
-                    }
-                },
                 {field: 'name'},
                 {field: 'groupName'},
                 {field: 'valueType'},
                 {field: 'value'},
                 {
                     field: 'overrideValue',
+                    omit: isNil,
                     formField: {
-                        className: 'xh-config-panel-override-value-field',
-                        label: false,
-                        flex: 1,
-                        inline: false,
-                        minimal: false,
-                        readonlyRenderer: v =>
-                            !isNil(v)
-                                ? hbox({
-                                      className: 'xh-config-panel-override-value-field__contents',
-                                      items: [
-                                          label({
-                                              className: 'xh-form-field-label',
-                                              item: 'Override Value'
-                                          }),
-                                          div({
-                                              className: 'xh-form-field-inner',
-                                              item: pre(v)
-                                          })
-                                      ]
-                                  })
-                                : null
+                        className: 'xh-bg-intent-warning',
+                        info: 'Overridden by Instance Config'
                     }
                 },
                 {field: 'note', formField: {item: textArea({height: 100})}},
@@ -192,7 +150,11 @@ export class ConfigPanelModel extends HoistModel {
             matchFields: ['name'],
             valueRenderer: v => {
                 if (isNil(v)) return '';
-                return v.valueType === 'pwd' ? '*****' : v.value;
+                return v.valueType === 'pwd'
+                    ? '*****'
+                    : !isNil(v.overrideValue)
+                      ? this.withOverrideWarning(v.value)
+                      : v.value;
             }
         });
     }
@@ -206,16 +168,14 @@ export class ConfigPanelModel extends HoistModel {
 
     private valueRenderer = (value, {record}) => {
         value = this.fmtValue(value, record);
-        if (!record.get('isOverridden')) return value;
-        return hbox({
-            alignItems: 'center',
-            className: 'xh-config-panel__overridden',
-            items: [Icon.warning({intent: 'warning'}), hspacer(), value]
-        });
+        if (isNil(record.get('overrideValue'))) return value;
+        return this.withOverrideWarning(value);
     };
 
     private valueTooltip = (value, {record}) =>
-        record.get('isOverridden') ? 'Overridden by Instance Config' : this.fmtValue(value, record);
+        !isNil(record.get('overrideValue'))
+            ? 'Overridden by Instance Config'
+            : this.fmtValue(value, record);
 
     private fmtValue(value, record) {
         switch (record.data.valueType) {
@@ -226,5 +186,16 @@ export class ConfigPanelModel extends HoistModel {
             default:
                 return value?.toString();
         }
+    }
+
+    private withOverrideWarning(value) {
+        return hbox({
+            alignItems: 'center',
+            items: [Icon.warning({intent: 'warning'}), hspacer(), value],
+            style: {
+                color: 'var(--xh-text-color-muted)',
+                textDecoration: 'line-through'
+            }
+        });
     }
 }